Fix parsing of value from Error response
The previous parsing code only parsed a number from _one_ char,
which works until the value reaches double digits and then breaks,
resulting in a horrendous amount of CAS retries which are all doomed to
fail because they _never_ get the correct current value of the counter.

The code now correctly parses the entire number, which reduced the
number of messages per operation by a lot.
Previously it was around 800 msgs-per-op, and now it's less than 10.

Signed-off-by: Linda Siemons <linda@catbrained.dev>