Fix issue with invalid analysis results due to stale reads
Reads from the sequentially consistent KV store might return stale
values. We can prevent this by first issuing a unique write, which
prevents the KV store from reordering our read in undesireable ways.
(Although, as far as I understand it, the store _could_ still reorder
the read without violating sequential consistency, but proving that
this reordering is legal would be prohibitively expensive so it doesn't
try to reorder. In any case, the Maelstrom provided seq-kv service
seems to behave in this way so we make use of that.)

Signed-off-by: Linda Siemons <linda@catbrained.dev>